Fixed Authentication and Improved Frontend
