--- project/dashboard/templates/dashboard/includes/accounts.html
@@ -1,7 +1,0 @@
-{% block content %}
-  <!-- Account -->
-  <div id="accounts" class="container-fluid">
-    <h1>Accounts</h1>
-    <div class="container" id="accounts-container"></div>
-  </div>
-{% endblock %}
