--- project/dashboard/templates/dashboard/includes/analytics.html
@@ -1,8 +1,0 @@
-{% block content %}
-  <!-- Analytics -->
-  <div id="analytics" class="container-fluid">
-    <h1>Analytics</h1>
-    <div class="container" id="analytics-container-instagram"></div>
-    <div class="container" id="analytics-container-twitter"></div>
-  </div>
-{% endblock %}
