--- project/dashboard/templates/dashboard/includes/settings.html
@@ -1,22 +1,0 @@
-{% block content %}
-  <!-- Settings -->
-  <div id="settings" class="container-fluid">
-    <h1>Settings</h1>
-    <div class="dashboard-row">
-      <div class="card">
-        <div class="container">
-          <p>iBlinkco Settings</p>
-          <form method="POST">
-            <p><input type="text" placeholder="iBlinkco Username" style="font-size:24px;"></p>
-            <p><input type="text" placeholder="iBlinkco Password" style="font-size:24px;"></p>
-          <br>
-          <p>Suggestion Preferences</p>
-            <p><input type="text" placeholder="iBlinkco Username" style="font-size:24px;"></p>
-            <p><input type="text" placeholder="iBlinkco Password" style="font-size:24px;"></p>
-            <p><textarea rows="8" cols="80"></textarea></p>
-          </form>
-        </div>
-      </div>
-    </div>
-  </div>
-{% endblock %}
