--- project/dashboard/templates/dashboard/includes/stream.html
@@ -1,8 +1,0 @@
-{% block content %}
-<div class="container-fluid" id="stream">
-	<h1>Stream</h1>
-	{% for follower in followers %}
-		<h2>{{ follower.name }}</h2>
-	{% endfor %}
-</div>
-{% endblock %}
