+++ project/static/js/dashboard/ui.js
@@ -1,50 +1,25 @@
-// showing dashboard row by default
-$("#admin").show();
+// Ajac Functions
+$(document).ready(function(){
+  $('#myModal').modal('show');
+  // Setup Form
+  $('#setup-connect-form').on('submit', function(event){
+    $.ajax({
+      type : 'POST',
+      url : '/setup-update',
+      data : {
+        website_name : $('#website-name').val(),
+        website_url : $('#website-url').val()
+      },
+      success: function (value) {
+        // var evalue = JSON.parse(value);
+        $('#setup-form-message').html(value)
+        if(value == "success"){
+          location.reload();
+        }
+      }
+    });
+    event.preventDefault();
+  });
 
-// hiding analytics by default
-$("#analytics").hide();
-$("#accounts").hide();
-$("#settings").hide();
- $("#stream").hide();
 
-// Hidding and Showing Dashboard content onclick
-
-$("#show-analytics").click(function(){
- $("#admin").hide();
- $("#analytics").show();
- $("#settings").hide();
- $("#accounts").hide();
- $("#stream").hide();
-});
-
-$("#show-accounts").click(function(){
-  $("#admin").hide();
-  $("#analytics").hide();
-  $("#settings").hide();
-  $("#accounts").show();
-  $("#stream").hide();
-});
-$("#show-home").click(function(){
-  $("#admin").show();
-  $("#analytics").hide();
-  $("#settings").hide();
-  $("#accounts").hide();
-  $("#stream").hide();
-
-});
-
-$("#show-settings").click(function(){
-  $("#admin").hide();
-  $("#analytics").hide();
-  $("#accounts").hide();
-  $("#settings").show();
-  $("#stream").hide();
-});
-
-$("#show-stream").click(function(){
-  $("#admin").hide();
-  $("#analytics").hide();
-  $("#accounts").hide();
-  $("#settings").hide();
-  $("#stream").show();
 });
