+++ project/users/views.py
@@ -13,6 +13,58 @@ import requests
 # Defining Blueprint var
 users = Blueprint('users', __name__, template_folder='templates')
 
+def creationFormating(returnedData):
+    print("Beggining Format")
+    # Defining User Varibles
+    email = returnedData['account']['email']
+    firstname = returnedData['account']['firstname']
+    lastname = returnedData['account']['lastname']
+    websiteName = returnedData['website']['website-name']
+    websiteUrl = returnedData['website']['website-url']
+
+    # Storing returned data
+    session['name'] = str(firstname) + " " + str(lastname)
+    session['email'] = email
+
+
+    # Checking if dictionaries exist
+
+    # Checking for complete setup
+    try:
+        setup_complete = returnedData['account']['setup_complete']
+        session['setup_complete'] = setup_complete
+    except Exception as e:
+        print("Setup incomplete")
+        print(e)
+
+    # Checking for twitter
+    try:
+        requestedTwitterUserData = ['description', 'friends_count', 'followers_count', 'location', 'name', 'screen_name']
+        returnedTwitterUserData = dict()
+
+        # Defining variables equal to json data
+        twitterData = returnedData['twitter']
+
+        for i in requestedTwitterUserData:
+            userItem = returnedData['twitter']['userData'][i]
+            returnedTwitterUserData[i] = userItem
+            print(userItem)
+        session['userTwitterData'] = returnedTwitterUserData
+        # session['twitter'] = twitterData
+    except Exception as e:
+        print('Twitter not connected')
+        print(e)
+
+    # Checking for instagram
+    try:
+        # Defining variables equal to json data
+        instagramData = returnedData['instagram']
+
+        session['instagram'] = instagramData
+    except Exception as e:
+        print('Instagram not connected')
+        print(e)
+
 # Register page url and function
 @users.route("/register", methods=['GET', 'POST'])
 def register():
@@ -27,31 +79,29 @@ def register():
         url = "http://localhost:8000/create-user"
         data = { 'firstname': form.firstname.data, 'lastname': form.lastname.data, 'email': form.email.data, 'password': form.password.data }
 
-        # Sending data to API
-        getData = requests.post(url = url, json = data) 
+        try:
+            # Sending data to API
+            getData = requests.post(url = url, json = data) 
 
-        # Getting returned data
-        returnedData = getData.json()
-        successMessage = returnedData[0]['message']
-        email = returnedData[0]['email']
-        firstname = returnedData[0]['firstname']
-        lastname = returnedData[0]['lastname']
+            # Getting returned data
+            returnedData = getData.json()
+            email = returnedData[0]['email']
+            firstname = returnedData[0]['firstname']
+            lastname = returnedData[0]['lastname']
 
-        # Defining variables equal to json data
-        instagramData = returnedData[1]
-        twitterData = returnedData[2]
-
-        # Storing returned data
-        session['name'] = str(firstname) + " " + str(lastname)
-        session['email'] = email
-        session['instagramData'] = instagramData
-        session['twitterData'] = twitterData
+            # Storing returned data
+            session['name'] = str(firstname) + " " + str(lastname)
+            session['email'] = email
+            session['instagramData'] = instagramData
+            session['twitterData'] = twitterData
 
 
-        # Alerting user account was created
-        flash(f'Account Created for {form.email.data} !', 'success')
-
-        return redirect(url_for('homepage.home'))
+            # Alerting user account was created
+            flash(f'Account Created for {form.email.data} !', 'success')
+            return redirect(url_for('homepage.home'))
+        except Exception as e:
+            print(e)
+            flash(f'Failed to Create Account')
 
     return render_template('users/register.html', title='Register', form=form)
 
@@ -73,52 +123,24 @@ def login():
         print(getData)
 
         # Getting returned data
-        returnedData = getData.json()
+        finalizedData = getData.json()
+        print(finalizedData[1])
+        session['user'] = finalizedData[1]
+        returnedData = finalizedData[0]
+        print(returnedData)
+
         try:
-            successMessage = returnedData['userdata']['message']
-            print(successMessage)
+
+            createFormat = creationFormating(returnedData)
+            session.permanent = True
+
+            return redirect(url_for('dashboard.home'))
+
         except Exception as e:
-            flash(f'Login Failed')
-            return redirect(url_for('users.login'))
+            print("e")
+            print(e)
+            flash(f'signin failed')
 
-        # Defining User Varibles
-        email = returnedData['userdata']['email']
-        firstname = returnedData['userdata']['firstname']
-        lastname = returnedData['userdata']['lastname']
-
-        # Defining variables equal to json data
-        instagramData = returnedData['returnedInfoInstagram']
-        twitterData = returnedData['returnedInfoTwitter']
-
-        # Storing returned data
-        session['name'] = str(firstname) + " " + str(lastname)
-        session['email'] = email
-        session['instagramData'] = instagramData
-        session['twitterData'] = twitterData 
-
-
-        # Storing all variables in session
-        session['instagramBio'] = instagramData['bio']
-        session['instagramUsername'] = instagramData['instagram_username']
-        session['instagramNumberOfFollowers'] = instagramData['number_of_followers']
-        session['instagramNumberOfFollowing'] = instagramData['number_of_following']
-        session['instagramNumberOfPost'] = instagramData['number_of_post']
-        session['instagramOnDesktop'] = instagramData['on_desktop']
-        session['instagramOnMobile'] = instagramData['on_mobile']
-        session['instagramOnWeb'] = instagramData['on_web']
-        
-        session['twitterBio'] = twitterData['bio']
-        session['twitterUsername'] = twitterData['twitter_username']
-        session['twitterNumberOfFollowers'] = twitterData['number_of_followers']
-        session['twitterNumberOfFollowing'] = twitterData['number_of_following']
-        session['twitterNumberOfPost'] = instagramData['number_of_post']
-        session['twitterOnDesktop'] = twitterData['on_desktop']
-        session['twitterOnMobile'] = twitterData['on_mobile']
-        session['twitterOnWeb'] = twitterData['on_web']
-
-        session.permanent = True
-
-        return redirect(url_for('homepage.home'))
     return render_template('users/login.html', title="Login", form=form)
 
 @users.route("/logout")
